refactor: Extract get-collection-choices (LPS-116692)
