+++ src/fragment-composition/index.ts
@@ -1,5 +1,3 @@
-import fs from 'fs';
-import glob from 'glob';
 import path from 'path';
 import voca from 'voca';
 
@@ -9,11 +7,10 @@ import {
   FRAGMENT_COLLECTION_SLUG_VAR,
   FRAGMENT_COMPOSITION_NAME_VAR,
   FRAGMENT_SLUG_VAR,
-  NEW_COLLECTION_MESSAGE,
-  NEW_COLLECTION_SHORT,
   NEW_COLLECTION_VALUE,
 } from '../utils/constants';
 import CustomGenerator from '../utils/custom-generator';
+import { getCollectionChoices } from '../utils/get-collection-choices';
 
 export default class FragmentCompositionGenerator extends CustomGenerator {
   async prompting(): Promise<void> {
@@ -36,7 +33,7 @@ export default class FragmentCompositionGenerator extends CustomGenerator {
       type: 'list',
       name: FRAGMENT_COLLECTION_SLUG_VAR,
       message: FRAGMENT_COLLECTION_SLUG_MESSAGE,
-      choices: this._getCollectionChoices(),
+      choices: getCollectionChoices(this.destinationRoot()),
       when: !this.hasValue(FRAGMENT_COLLECTION_SLUG_VAR),
     });
   }
@@ -69,41 +66,5 @@ export default class FragmentCompositionGenerator extends CustomGenerator {
         'definition.json',
       ]);
     }
-  }
-
-  private _getCollectionChoices() {
-    let choices: Array<{
-      name: string;
-      value: string;
-      short: string;
-    }> = [];
-
-    try {
-      choices = glob
-        .sync(`${this.destinationRoot()}/src/*/collection.json`)
-        .map((collectionJSON) => {
-          const collectionName = JSON.parse(
-            fs.readFileSync(collectionJSON, 'utf-8')
-          ).name;
-
-          const collectionSlug = path.basename(
-            path.resolve(`${collectionJSON}/..`)
-          );
-
-          return {
-            name: collectionName,
-            value: collectionSlug,
-            short: `(${collectionSlug})`,
-          };
-        });
-    } catch (_) {}
-
-    choices.push({
-      name: NEW_COLLECTION_MESSAGE,
-      value: NEW_COLLECTION_VALUE,
-      short: NEW_COLLECTION_SHORT,
-    });
-
-    return choices;
   }
 }
